# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reco API (mock)",
    "version": "v1"
  },
  "paths": {
...
    "/v1/recommend": {
      "post": {
        "tags": [
          "recommend"
        ],
        "summary": "Recommend",
        "operationId": "recommend_v1_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recommendations/{request_id}": {
      "get": {
        "tags": [
          "recommend"
        ],
        "summary": "Get Log",
        "operationId": "get_log_v1_recommendations__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankedItem": {
        "properties": {
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id"
          },
          "missing_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Missing Count"
          },
          "missing_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Missing Ids"
          },
          "coverage": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Coverage"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "features": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Features",
            "description": "피처 원값 17종. w 와 무관하게 항상 전부 기록한다"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "penalty": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Penalty",
            "description": "p_recent · p_cooked · (1-p_avoid) 의 곱",
            "default": 1.0
          },
          "final_rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Final Rank"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "템플릿 생성 문구 (설계 5-5)",
            "default": ""
          },
          "reason_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Features",
            "description": "이유를 만든 피처 (z-salience 상위). 디버거가 근거를 보여준다"
          },
          "mmr_penalty": {
            "type": "number",
            "title": "Mmr Penalty",
            "default": 0.0
          },
          "is_exploration": {
            "type": "boolean",
            "title": "Is Exploration",
            "description": "무작위 삽입 슬롯. position bias 보정의 기준점 (설계 5-3-3)",
            "default": false
          },
          "propensity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Propensity",
            "description": "노출 확률. exploration 슬롯은 1/|pool|, 결정적 슬롯은 1.0"
          },
          "explore_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explore Source",
            "description": "uniform | thompson | None(탐색 슬롯이 아님)"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recipe_id",
          "missing_count",
          "coverage",
          "features",
          "score",
          "final_rank"
        ],
        "title": "RankedItem",
        "description": "③ 산출. 유저에게 나가는 최종 형태."
      },
...
      "RecommendRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[cgd]-"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "클라이언트 발급 (c-{user}-{uuid12}). 30분 무활동 시 갱신"
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 20
          },
          "max_missing": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Max Missing",
            "default": 2
          },
          "max_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Minutes"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "weight_override": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Override"
          },
          "interleave_with": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interleave With",
            "description": "비교 대상 model_version. 지정 시 items[].team 이 채워진다"
          },
          "include_trace": {
            "type": "boolean",
            "title": "Include Trace",
            "description": "False 면 응답에서 trace 를 빼고 DB 에는 그대로 남긴다",
            "default": true
          },
          "context": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Context",
            "description": "{hour, weekday, device, source_screen}"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "RecommendRequest"
      },
      "RecommendResponse": {
        "properties": {
          "contract_version": {
            "type": "string",
            "title": "Contract Version",
            "default": "v1"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id",
            "description": "event 기록 시 이 값을 함께 보낸다"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RankedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "served_at": {
            "type": "string",
            "format": "date-time",
            "title": "Served At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "user_id",
          "model_version",
          "items",
          "served_at"
        ],
        "title": "RecommendResponse"
      },
      "RecommendationLogOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[cgd]-"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "mlflow_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mlflow Run Id"
          },
          "config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Hash"
          },
          "warm_alpha": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Warm Alpha"
          },
          "stats_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats Version"
          },
          "pantry_snapshot": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pantry Snapshot"
          },
          "pantry_detail": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pantry Detail"
          },
          "allergy_snapshot": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Allergy Snapshot"
          },
          "policies": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policies"
          },
          "stage_trace": {
            "$ref": "#/components/schemas/StageTrace"
          },
          "served": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Served"
          },
          "total_latency_ms": {
            "type": "integer",
            "title": "Total Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "user_id",
          "model_version",
          "pantry_snapshot",
          "stage_trace",
          "served",
          "total_latency_ms",
          "created_at"
        ],
        "title": "RecommendationLogOut",
        "description": "`recommendation_log` 1행. **DDL 과 필드가 일치해야 한다** (v2.9 동기화).\n\nv2.8 DDL 개정으로 컬럼 7개가 늘었는데 이 계약이 따라오지 않아\n**DB 에 칸은 있는데 채울 계약이 없는** 상태였다. 여기서 맞춘다."
      },
      "Stage": {
        "type": "string",
        "enum": [
          "retrieval",
          "ranking",
          "rerank",
          "mealplan"
        ],
        "title": "Stage"
      },
      "StageInfo": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/Stage"
          },
          "in_count": {
            "type": "integer",
            "title": "In Count"
          },
          "out_count": {
            "type": "integer",
            "title": "Out Count"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "fallback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback",
            "description": "폴백이 발동했으면 대체 모델명 (설계 5-6)"
          },
          "filters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Filters",
            "description": "탈락 사유별 건수. 예 {'no_overlap':227999,'allergy_cut':12}"
          },
          "dropped": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Dropped",
            "description": "③ 전용. 다양성·캡으로 제외한 건수"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Params"
          },
          "score_stats": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Score Stats",
            "description": "min·p25·p50·p75·max"
          },
          "exploration_items": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Exploration Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "in_count",
          "out_count",
          "latency_ms"
        ],
        "title": "StageInfo",
        "description": "단계 하나의 기록. filters 가 디버깅에서 가장 유용하다."
      },
      "StageTrace": {
        "properties": {
          "trace_version": {
            "type": "string",
            "title": "Trace Version",
            "default": "v1"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageInfo"
            },
            "type": "array",
            "title": "Stages"
          },
          "totals": {
            "$ref": "#/components/schemas/TraceTotals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "totals"
        ],
        "title": "StageTrace",
        "description": "recommendation_log.stage_trace 에 그대로 직렬화된다."
      },
      "TraceTotals": {
        "properties": {
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "description": "폴백 경로를 탔다. 이 비율이 조용히 오르는 것이 가장 위험한 실패 양상",
            "default": false
          },
          "user_mode": {
            "$ref": "#/components/schemas/UserMode",
            "default": "onboarding"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latency_ms"
        ],
        "title": "TraceTotals"
      },
      "UserMode": {
        "type": "string",
        "enum": [
          "onboarding",
          "blended",
          "behavior"
        ],
        "title": "UserMode",
        "description": "user_vector.computed_from 과 동일한 값 (설계 2-5)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}